+++ lightning/api_clnsocket.go
