+++ lightning/api_clnsocket_test.go
