+++ lightning/cln_entities.go
