+++ lightning/fixtures/cln_info.json.template
