+++ lightning/fixtures/cln_listchans.json.template
