+++ lightning/fixtures/getinfo_grpc.json
