+++ lightning/fixtures/graph_edge_grpc.json
