+++ lightning/fixtures/graph_node.json
