+++ lightning/fixtures/graph_node_grpc.json
